fix: rename api-requets to api-request
